refactor(sessions): drop cross-workspace session relations
Cross-worktree tracked subsessions are no longer possible, so listings no
longer annotate sessions with relations that live in a sibling workspace.
Removes the sibling-workspace child count (a reverse-index query with no
index, which listed every sibling workspace in full on every listing) and
the out-of-listing parent-location lookup.

The shared SessionInfo fields (parentSessionCwd, parentSessionId,
childSessionsElsewhere) stay for now; they are removed together with the
client UI that reads them.

parentSessionPath, same-worktree session trees, detach memo invalidation,
sessionFileHeader.ts, and the spawn target resolver are unaffected.
