+++ src/server/sessions/piSessionService.ts
@@ -62,8 +62,6 @@ import type { SessionRouteRef, SessionRouteService } from "./sessionService.js";
 import { type AuthChange } from "./authService.js";
 import { canonicalizeStoredCwd, cwdPathsEqual } from "../workingDirectory.js";
 import { readSessionHeaderSummary } from "./sessionFileHeader.js";
-import { countOutOfListingChildren, locateOutOfListingParents } from "./parentSessionLocator.js";
-import { siblingWorkspaceCwds, type ProjectWorkspaceCwds } from "../workspaces/projectWorkspaceCwds.js";
 import type { WorkspaceActivityService } from "../activity/workspaceActivityService.js";
 import { createAskUserToolDefinition, type AskUserInvocation, type AskUserToolDeps } from "./askUserTool.js";
 import { PendingAskStore, renderAskUserAnswersText, type PendingAskCloseResult, type PendingAskOpenResult } from "./pendingAskStore.js";
@@ -761,13 +759,6 @@ export interface PiSessionServiceDependencies {
    */
   spawnTargets?: SpawnTargetResolver;
   /**
-   * When provided, session listings report related sessions living in sibling
-   * workspaces of the same project: where an out-of-workspace parent is, and how
-   * many children a listed session has elsewhere. Omit to list each workspace in
-   * isolation.
-   */
-  projectWorkspaces?: ProjectWorkspaceCwds;
-  /**
    * Beta: when true (and `spawnTargets` is provided), the tracked-subsession
    * tools are available to sessions whose creation provenance permits
    * delegation. Off by default so the capability can ship in main without
@@ -855,7 +846,6 @@ export class PiSessionService implements SessionRouteService {
   private readonly modelRuntime: ModelRuntime;
   private readonly workspaceActivity: Pick<WorkspaceActivityService, "applySessionStatus" | "applySessionActivity" | "removeSession" | "reconcileSessionActivity"> | undefined;
   private readonly spawnTargets: SpawnTargetResolver | undefined;
-  private readonly projectWorkspaces: ProjectWorkspaceCwds | undefined;
   private readonly logger: PiSessionLogger;
   private readonly now: () => Date;
   private readonly notificationStore: SessionNotificationStore;
@@ -882,7 +872,6 @@ export class PiSessionService implements SessionRouteService {
     this.sessionManager = deps.sessionManager;
     this.modelRuntime = deps.modelRuntime;
     this.spawnTargets = deps.spawnTargets;
-    this.projectWorkspaces = deps.projectWorkspaces;
     this.logger = deps.logger ?? noopLogger;
     this.now = deps.now ?? (() => new Date());
     this.notificationStore = deps.notificationStore ?? new SessionNotificationStore();
@@ -1105,63 +1094,7 @@ export class PiSessionService implements SessionRouteService {
       .sort(compareArchivedRecords)
       .map((record) => clientSessionFromArchivedRecord(record, sessionsById.get(record.sessionId)))
       .filter(isDefined);
-    return await this.withRelatedSessionsElsewhere([...unarchivedSessions, ...archivedSessions], cwd);
-  }
-
-  /**
-   * Annotate a listing with the session relationships that cross workspace
-   * boundaries: where an out-of-listing parent lives, and how many children a
-   * listed session has in sibling workspaces.
-   *
-   * Both directions are best-effort. An unreadable parent header or a sibling
-   * workspace that cannot be listed leaves the session unannotated rather than
-   * failing the listing, and the browser falls back to its generic states.
-   */
-  private async withRelatedSessionsElsewhere(sessions: readonly ClientSession[], cwd: string): Promise<ClientSession[]> {
-    const [parentLocations, childCounts] = await Promise.all([
-      locateOutOfListingParents(sessions, cwd, readSessionHeaderSummary),
-      this.countChildrenInSiblingWorkspaces(sessions, cwd),
-    ]);
-    return sessions.map((session) => {
-      const parent = session.parentSessionPath === undefined ? undefined : parentLocations.get(session.parentSessionPath);
-      const childrenElsewhere = childCounts.get(session.path);
-      if (parent === undefined && childrenElsewhere === undefined) return session;
-      const annotated = { ...session };
-      if (parent !== undefined) {
-        annotated.parentSessionId = parent.parentSessionId;
-        annotated.parentSessionCwd = parent.parentSessionCwd;
-      }
-      if (childrenElsewhere !== undefined) annotated.childSessionsElsewhere = childrenElsewhere;
-      return annotated;
-    });
-  }
-
-  /**
-   * Count children of the listed sessions that live in other workspaces of the
-   * same project.
-   *
-   * Only sibling workspaces are scanned: agents may only spawn into workspaces
-   * of the spawning session's own project, so that bounds where a child can be.
-   * Listing is skipped entirely when no project-workspace locator is configured
-   * or the cwd belongs to no registered project.
-   */
-  private async countChildrenInSiblingWorkspaces(sessions: readonly ClientSession[], cwd: string): Promise<Map<string, number>> {
-    if (this.projectWorkspaces === undefined || sessions.length === 0) return new Map();
-    try {
-      const siblingCwds = await siblingWorkspaceCwds(this.projectWorkspaces, cwd);
-      if (siblingCwds.length === 0) return new Map();
-      const listings = await Promise.all(siblingCwds.map(async (siblingCwd): Promise<string[]> => {
-        const entries = await this.sessionManager.list(siblingCwd);
-        return entries.flatMap((entry) => entry.parentSessionPath === undefined ? [] : [entry.parentSessionPath]);
-      }));
-      return countOutOfListingChildren(sessions, listings.flat());
-    } catch (error: unknown) {
-      this.logger.info(
-        { cwd, error: error instanceof Error ? error.message : String(error) },
-        "failed to count child sessions in sibling workspaces",
-      );
-      return new Map();
-    }
+    return [...unarchivedSessions, ...archivedSessions];
   }
 
   async start(cwd: string, options: StartSessionOptions = {}): Promise<ClientSession> {
